+++ hms-api/src/rest/IdeaResponse.ts
@@ -23,7 +23,6 @@ class IdeaResponse {
   goal: string;
   requiredSkills: SkillPreviewResponse[];
   category: CategoryPreviewResponse;
-  finalVideoUrl: string;
   creationDate: Date;
 
   constructor(
@@ -37,7 +36,6 @@ class IdeaResponse {
     goal: string,
     requiredSkills: SkillPreviewResponse[],
     category: CategoryPreviewResponse,
-    finalVideoUrl: string,
     creationDate: Date,
   ) {
     this.id = id;
@@ -50,7 +48,6 @@ class IdeaResponse {
     this.goal = goal;
     this.requiredSkills = requiredSkills;
     this.category = category;
-    this.finalVideoUrl = finalVideoUrl;
     this.creationDate = creationDate;
   }
 
@@ -75,7 +72,6 @@ class IdeaResponse {
       idea.goal,
       SkillPreviewResponse.fromArray(skills),
       CategoryPreviewResponse.from(category),
-      idea.finalVideoUrl,
       idea.creationDate,
     );
 }
