+++ hms-api/src/service/idea-service.ts
@@ -44,7 +44,6 @@ export async function createIdea(
   goal: string,
   requiredSkills: Uuid[],
   categoryId: Uuid,
-  finalVideoUrl: string,
 ): Promise<Idea> {
   if (!(await participantExistsForHackathon(ownerId, hackathonId))) {
     throw new ReferenceNotFoundError(
@@ -82,7 +81,6 @@ export async function createIdea(
     goal,
     requiredSkills,
     categoryId,
-    finalVideoUrl,
   );
   const result = idea.validate();
   if (result.hasFailed()) {
@@ -101,7 +99,6 @@ export async function editIdea(
   goal: string,
   requiredSkills: Uuid[],
   categoryId: Uuid,
-  finalVideoUrl: string,
 ): Promise<void> {
   let existing: Idea;
   try {
@@ -135,7 +132,6 @@ export async function editIdea(
   existing.goal = goal;
   existing.requiredSkills = requiredSkills;
   existing.categoryId = categoryId;
-  existing.finalVideoUrl = finalVideoUrl;
 
   const result = existing.validate();
   if (result.hasFailed()) {
@@ -357,8 +353,4 @@ export async function removeParticipantFromIdeas(
       );
     }
   }
-}
-
-export async function createFinalVideo(): Promise<void> {
-  // TODO
 }
